Delete 'eventlet.monkey_patch()' on 'app' module.
WHY this dangerous method (which totally patches some Python's internal
module!) was introduced to app/__init__.py ? It'is a long story...

At the beginning, we didn't seperate `add` and `websocket server`,
they're working on the same process, then I found if acquiring
`websocket server` with message queue, it will block the whole HTTP
request! Thus, in order to solve this problem, I introduced it to patch
'threading' module.

However, in `ws_alone` branch (I merged one month ago), I have
successfully seperated it to a standalone process to solve the blocking
problem eventually. So the monkey_patch method was no longer necessary.

But I forgot to delete it then... Until I take a lot of time to find the
source the problem!

This story told us that: monkey_patch is really DANGERAOUS!! You should
be carefull about that!

Fix #2